main_pipeline/requirements.txt
biopython>=1.81
tqdm>=4.64.0
python-dotenv>=1.0.0 
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.7.1
google-auth>=2.15.0

# Новые зависимости для LCGR
instructor>=1.0.0
google-genai>=0.8.0
networkx>=3.0
matplotlib>=3.0.0
pydantic>=2.0.0
scikit-learn>=1.0.0
arxiv>=2.0.0
requests>=2.28.0 